Refactor JSON schema handling: move schema definitions to separate file
